ISPN-12991 Move the implementation on different class
In this way the core functions may work without the dependency of Micrometer, thus allowing the optionality of the dependency
