+++ core/src/main/java/org/infinispan/metrics/impl/MetricsCollectorFactory.java
@@ -10,11 +10,6 @@ import org.infinispan.factories.scopes.Scopes;
 import org.infinispan.util.logging.Log;
 import org.infinispan.util.logging.LogFactory;
 
-import io.micrometer.core.instrument.Meter;
-import io.micrometer.core.instrument.config.MeterFilter;
-import io.micrometer.prometheus.PrometheusConfig;
-import io.micrometer.prometheus.PrometheusMeterRegistry;
-
 /**
  * Produces instances of {@link MetricsCollector}. MetricsCollector is optional,
  * based on the presence of Micrometer in classpath and the enabling of metrics in config.
@@ -40,43 +35,11 @@ public final class MetricsCollectorFactory implements ComponentFactory, AutoInst
 
       // try cautiously
       try {
-         return create();
+         return new MetricsCollector();
       } catch (Throwable e) {
          // missing dependency
          log.debug("Micrometer metrics are not available because classpath dependencies are missing.", e);
          return null;
-      }
-   }
-
-   private MetricsCollector create() {
-      PrometheusMeterRegistry baseRegistry = new PrometheusMeterRegistry(PrometheusConfig.DEFAULT);
-      baseRegistry.config().meterFilter(new BaseFilter());
-
-      new BaseAdditionalMetrics().bindTo(baseRegistry);
-
-      PrometheusMeterRegistry vendorRegistry = new PrometheusMeterRegistry(PrometheusConfig.DEFAULT);
-      vendorRegistry.config().meterFilter(new VendorFilter());
-
-      new VendorAdditionalMetrics().bindTo(vendorRegistry);
-
-      return new MetricsCollector(baseRegistry, vendorRegistry);
-   }
-
-   private static class BaseFilter implements MeterFilter {
-      private static final String PREFIX = "base.";
-
-      @Override
-      public Meter.Id map(Meter.Id id) {
-         return id.withName(PREFIX + id.getName());
-      }
-   }
-
-   private static class VendorFilter implements MeterFilter {
-      private static final String PREFIX = "vendor.";
-
-      @Override
-      public Meter.Id map(Meter.Id id) {
-         return id.withName(PREFIX + id.getName());
       }
    }
 }
